Some fixes to be same as PyTorch version

+++ scannet/scannet_detection_dataset.py
@@ -320,7 +320,7 @@ class ScannetDetectionDataset(Dataset):
         target_bboxes_semcls[0:instance_bboxes.shape[0]] = \
             [DC.nyu40id2class[x] for x in instance_bboxes[:,-1][0:instance_bboxes.shape[0]]]                
         ret_dict['sem_cls_label'] = target_bboxes_semcls.astype(np.int64)
-        ret_dict['box_label_mask'] = target_bboxes_mask.astype(np.float32)
+        ret_dict['box_label_mask'] = target_bboxes_mask.astype(np.int64)
         ret_dict['point_obj_mask'] = point_obj_mask.astype(np.int64)
         ret_dict['point_instance_label'] = point_instance_label.astype(np.int64)
         ret_dict['size_gts'] = size_gts.astype(np.float32)
